refactor: data 모듈 패키지 생성 및 정리
