+++ data/src/main/java/com/github/dodobest/data/remote/UpbitAPI.kt
@@ -1,5 +1,7 @@
-package com.github.dodobest.data
+package com.github.dodobest.data.remote
 
+import com.github.dodobest.data.model.UpbitMarketData
+import com.github.dodobest.data.model.UpbitTickerData
 import io.reactivex.rxjava3.core.Single
 import retrofit2.http.GET
 import retrofit2.http.Query
