+++ data/src/main/java/com/github/dodobest/data/remote/UpbitRemoteDataSourceImpl.kt
@@ -1,10 +1,12 @@
-package com.github.dodobest.data
+package com.github.dodobest.data.remote
 
+import com.github.dodobest.data.model.UpbitMarketData
+import com.github.dodobest.data.model.UpbitTickerData
 import io.reactivex.rxjava3.core.Single
 
 internal class UpbitRemoteDataSourceImpl(
     private val upbitAPI: UpbitAPI
-) : UpbitRemoteDataSource{
+) : UpbitRemoteDataSource {
     override fun getMarkets(): Single<List<UpbitMarketData>> {
         return upbitAPI.getMarkets()
     }
